Сделала задание Помощь друга
